pyproject.toml
[project]
name = "document-intelligence-api"
version = "1.0.0"
description = "Intelligent document analysis API with context-aware extraction"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "langextract",
    "google-generativeai>=0.8.3",
    "pdfplumber>=0.11.4",
    "python-docx>=1.1.2",
    "reportlab>=4.2.5",
    "python-dotenv>=1.0.1",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
